+++ src/duplocloud/args.py
@@ -301,6 +301,31 @@ TARGETS = Arg("targets", "--targets",
 TIME = Arg("time", "--time",
            help='The time to use')
 
+TITLE = Arg("title", "--title",
+            help= "The Title for ticket",
+            required=True)
+
+AGENTNAME = Arg("agent_name", "--agent_name", "--agent",
+                help= "AI Agent to be used to process the ticket",
+                required=True)
+
+INSTANCEID = Arg("instance_id","--instance_id", "--instance",
+                help= "AI Agent Instance Id",
+                required=True)
+
+APIVERSION = Arg("api_version", "--api-version",
+                help="API Version",
+                required=False,
+                default="v1")
+
+TICKETID = Arg("ticket_id", "--ticket_id", "--ticket",
+              help="The ID of the AI HelpDesk ticket",
+              required=True)
+
+MESSAGE = Arg("content", "--content", "--msg",
+              help="The message you want to send to the AI agent",
+              required=True)
+
 DAYS = Arg("days", 
             help='The days to use',
             type=int)